backend/pyproject.toml
[tool.poetry]
name = "online-cinema-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
aiohttp = "^3.11.9"
aiobotocore = "^2.15.2"
numpy = "<2"
uvicorn = "^0.34.0"
aioboto3 = "^13.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
scikit-learn = "^1.6.0"
implicit = "^0.7.2"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
jupyter = "^1.1.1"
openpyxl = "^3.1.5"
plotly = "^5.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
